Demonstrate better behavior with compiledBy(...)
